requirements.txt
# Core & Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.13.1
pydantic>=2.5.1
pydantic-settings>=2.1.0

# File & Config Management
watchdog>=3.0.0
PyYAML>=6.0.1
python-frontmatter>=1.0.0
markdown>=3.5.2

# LLM & NLP (compatíveis com Python 3.13)
llama-cpp-python>=0.2.26
sentence-transformers>=2.2.2
transformers>=4.36.2
# torch será instalado via sistema ou conda
safetensors>=0.4.2
langchain>=0.1.0
langchain-community>=0.0.10

# Text Processing
spacy>=3.7.2
nltk>=3.8.1
pymupdf>=1.23.8
python-docx>=1.1.0
ebooklib>=0.18
beautifulsoup4>=4.12.2
lxml>=4.9.3

# CLI & Interface
typer>=0.9.0
rich>=13.7.0
questionary>=2.0.1
textual>=0.41.0
prompt-toolkit>=3.0.43

# Data & Utilities
pandas>=2.2.0
numpy>=1.26.3
matplotlib>=3.8.2
seaborn>=0.13.0
python-dateutil>=2.8.2
icalendar>=5.0.11
pytz>=2023.3.post1
aiosqlite>=0.19.0
redis>=5.0.1

# Optional (comment if not needed)
# gtts>=2.4.0
# playsound>=1.3.0
# pydub>=0.25.1
# googletrans>=4.0.0-rc1
# deep-translator>=1.11.4
